final_version/requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
aiohttp=3.6.3=py37h7b6447c_0
argon2-cffi=20.1.0=py37h7b6447c_1
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=py37_0
async_generator=1.10=py37h28b3542_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37h7b6447c_1000
bzip2=1.0.8=h7b6447c_0
c-ares=1.16.1=h7b6447c_0
ca-certificates=2020.10.14=0
cachetools=4.1.1=py_0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=py37_1003
chumpy=0.70=pypi_0
click=7.1.2=py_0
cloudpickle=1.5.0=pypi_0
cryptography=3.1.1=py37h1ba5d50_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
detectron2=0.2.1+cu102=pypi_0
easydict=1.9=pypi_0
entrypoints=0.3=py37_0
expat=2.2.10=he6710b0_2
ffmpeg=1.4=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.3=h5ab3b9f_0
freetype-py=2.2.0=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.1.post20200716=py_0
gast=0.3.3=pypi_0
glib=2.66.1=h92f7085_0
google-auth=1.4.2=pypi_0
google-auth-oauthlib=0.4.1=py_2
google-colab=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=h23475e2_0
grpcio=1.30.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
idna=2.8=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=4.6.1=pypi_0
ipython=5.5.0=pypi_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_1
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.16.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py37_0
jupyterlab_pygments=0.1.2=py_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py37hfd86e86_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=heb0550a_3
markdown=3.3.1=py37_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.0=pypi_0
matplotlib-base=3.3.1=py37h817c723_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mock=4.0.2=pypi_0
msgpack=1.0.0=pypi_0
multidict=4.7.6=py37h7b6447c_1
natsort=7.0.1=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.1=py_0
networkx=2.4=pypi_0
ninja=1.10.1=py37hfd86e86_0
notebook=5.2.2=pypi_0
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
open3d=0.9.0.0=pypi_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=4.3.0.36=pypi_0
openssl=1.1.1h=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=py_0
pandas=0.24.2=pypi_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pascal-voc-writer=0.1.4=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py37_1
pickleshare=0.7.5=py37_1001
pillow=6.1.0=pypi_0
pip=19.0=pypi_0
pixman=0.40.0=h7b6447c_0
portalocker=2.0.0=py37_0
portpicker=1.2.0=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=1.0.18=pypi_0
prompt_toolkit=3.0.8=0
protobuf=3.12.2=pypi_0
ptyprocess=0.6.0=py37_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.1=pypi_0
pycparser=2.20=py_2
pydot=1.4.1=pypi_0
pyglet=1.5.7=pypi_0
pygments=2.7.1=py_0
pyjwt=1.7.1=py37_0
pyopengl=3.1.0=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pyqt5=5.15.0=pypi_0
pyqt5-sip=12.8.1=pypi_0
pyqtgraph=0.11.0=pypi_0
pyrender=0.1.43=pypi_0
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.6.0=py3.7_cuda10.2.89_cudnn7.6.5_0
pytorch3d=0.2.5=py37_cu102_pyt160
pytz=2020.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=19.0.2=py37he6710b0_1
qt=5.9.7=h5867ecd_1
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
quaternion=2020.5.19.15.27.24=py37h03ebfcd_0
quaternions=0.4=pypi_0
readline=8.0=h7b6447c_0
requests=2.21.0=pypi_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scandir=1.10.0=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.1.0=pypi_0
send2trash=1.5.0=py37_0
setuptools=50.3.0=py37hb0f4dca_1
simplegeneric=0.8.1=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.12.0=pypi_0
smplx=0.1.15=pypi_0
sqlite=3.33.0=h62c20be_0
tabulate=0.8.7=py37_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow-estimator=2.2.0=pypi_0
tensorflow-gpu=2.2.0=pypi_0
termcolor=1.1.0=py37_1
terminado=0.9.1=py37_0
terminaltables=3.1.0=pypi_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.7.4=pypi_0
tk=8.6.10=hbc83047_0
torch=1.6.0=pypi_0
torchgeometry=0.1.2=pypi_0
torchvision=0.7.0=pypi_0
tornado=4.5.3=pypi_0
tqdm=4.50.2=py_0
traitlets=5.0.5=py_0
trimesh=3.7.12=pypi_0
urllib3=1.24.3=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.2=py37h7b6447c_0
youtube-dl=2020.6.16.1=pypi_0
zeromq=4.3.3=he6710b0_3
zipp=3.3.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
